+++ example/ios/RunnerTests/RunnerTests.swift
@@ -3,7 +3,7 @@ import UIKit
 import XCTest
 
 
-@testable import zebra_scanner
+@testable import zebra_handheld_scanner
 
 // This demonstrates a simple unit test of the Swift portion of this plugin's implementation.
 //